+++ src/ext/java/org/opentripplanner/ext/legacygraphqlapi/datafetchers/LegacyGraphQLTripImpl.java
@@ -170,7 +170,7 @@ public class LegacyGraphQLTripImpl implements LegacyGraphQLDataFetchers.LegacyGr
         );
 
         Stop stop = timetable.getPattern().getStop(0);
-        return new TripTimeOnDate(triptimes, 0, stop, serviceDate
+        return new TripTimeOnDate(triptimes, 0, tripPattern, serviceDate
         );
       } catch (ParseException e) {
         //Invalid date format
@@ -201,7 +201,7 @@ public class LegacyGraphQLTripImpl implements LegacyGraphQLDataFetchers.LegacyGr
           );
 
         Stop stop = timetable.getPattern().getStop(triptimes.getNumStops() - 1);
-        return new TripTimeOnDate(triptimes, triptimes.getNumStops() - 1, stop, serviceDate
+        return new TripTimeOnDate(triptimes, triptimes.getNumStops() - 1, tripPattern, serviceDate
         );
       } catch (ParseException e) {
         //Invalid date format